requirements.txt
asgiref==3.5.2
Django==4.1.4
sqlparse==0.4.3
tzdata==2022.7asgiref==3.5.2
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
asgiref==3.5.2
attrs==22.1.0
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.1
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
